requirements.txt
flask~=2.2.2
requests~=2.31.0
transformers~=4.29.2
numpy~=1.24.3
torch==2.2.1
pillow~=9.4.0

omegaconf==2.2.3
pandas==1.5.2
#pytorch_lightning==1.8.3.post1
torchaudio==2.2.1
tqdm==4.64.1
webdataset==0.2.31
python-dotenv==0.21.1
torchvision==0.17.1
wandb==0.13.9
open-clip-torch==2.20.0
scikit-learn==1.3.0
prometheus-client==0.17.0
codecarbon==2.2.4
peft==0.5.0
